% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoJFMP.R
\name{autoJFMP}
\alias{autoJFMP}
\title{Create Automatic JFMP burn allocation}
\usage{
autoJFMP(myJFMP1, myTargetHa)
}
\arguments{
\item{myJFMP1}{data frame of planning unit attributes and JFMP scores for each polygon burned and no t burned between JFMPSEASON0 and JFMPSEASON 0 +4 as produced by function jfmp1()}

\item{myTargetHa}{data frame of target areas for each Fire District and FMZ_CODE in the study area in Hectares}
}
\value{
JFMP auto-nominations	 table 1, with one row for each planning unit, and columns for:
– Planning unit ID (x1 column)
– FMZ_CODE category (x1)
– District (x1)
– Planning unit size in hectares (x1)
above columns are sourced from the input planning unit shapefile.
– Score (now Diff) for each metric in burn/non-burn states (2 x 4)
– Difference in each metric between burn/non-burn states (x4)
– Ranking on difference in each metric between states within District (x4)
– Ranking as above but within District X FMZ_CODE combination (x 4)
}
\description{
Create Automatic JFMP burn allocation
}
