% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_DraftSpList.R
\name{calc_DraftSpList}
\alias{calc_DraftSpList}
\title{Calculate a draft species list for defined polygon}
\usage{
calc_DraftSpList(
  REG_NO,
  RasterRes = 225,
  PUBLIC_LAND_ONLY,
  myPoly = clipPoly,
  generalRasterDir = "./InputGeneralRasters",
  TaxonList = "./ReferenceTables/DraftTaxonListStatewidev2.csv"
)
}
\arguments{
\item{REG_NO}{integer DELWP fire region number 1:6 ,99 for Statewide analysis, or 7 for ad-hoc boundary polygon default =7 (see look up table REG_LUT for values)}

\item{RasterRes}{integer 225 - raster resolution is always 225 for this function for speed}

\item{PUBLIC_LAND_ONLY}{logical whether to restrict analysis to public land only or the whole polygon}

\item{myPoly}{default clipPoly sf polygon data frame of LF_REGIONs (default) or ad hoc polygon - used in conjunction with REG_NO}

\item{generalRasterDir}{relative path to directory containing rasters of FIRE_REG, and PUBLIC LAND (PLM_GEN)}

\item{TaxonList}{path to default species attribute table coantianing paths to HDMs in column HDMPath}
}
\value{
data.frame created from  TaxonList with columns appended for:
\itemize{
\item cellsInState count of the number of cells in the state within the Binary HDM for the species
\item cellsInArea count of the number of cells within myPoly and within the Binary HDM for the species
\item areaProp proportion of binary HDM for the state within myPoly
}
}
\description{
Calculate a draft species list for defined polygon
}
\details{
Calculate the proportion of cells for the HDM in the region for each species
is intended only as a starting point and requires manual quality control to
produce a useful species list for the area by editing the resulting .csv
file
}
