% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_SpeciesRA.R
\name{calc_SpeciesRA}
\alias{calc_SpeciesRA}
\title{Species' relative abundance calculation and summary}
\usage{
calc_SpeciesRA(
  myFHAnalysis,
  myAllCombs = allCombs,
  myHDMSpp_NO = HDMSpp_NO,
  myWriteSpRasters = TRUE,
  myLU_List = LU_List,
  myResultsDir = ResultsDir,
  myTaxonList = TaxonList,
  writeYears = NULL,
  myWriteSp = writeSp,
  myCropRasters = cropRasters
)
}
\arguments{
\item{myFHAnalysis}{list containing all the fire history spatial attributes
created by function fhProcess}

\item{myAllCombs}{all combinations of raster values object produced by
function calc_U_AllCombs}

\item{myHDMSpp_NO}{vector of TAXON_IDs for species to be included in output}

\item{myWriteSpRasters}{logical: whether to also write species abundance
rasters to disk}

\item{myLU_List}{list of species abundance lookup arrays created by function
make_Spp_LU_list()}

\item{myResultsDir}{path of directory where raster results will be written usually
generated  by FAME script}

\item{myTaxonList}{data.frame of species attributes (read from default or
user provided .csv) including field HDMPaths with the relative or absolute
path to the HDM raster used in the analysis.}

\item{writeYears}{vector for SEASONS for which rasters are to be written
otherwise if writeSpRasters == TRUE, if writeYears == NULL then all SEASONS
are written out}

\item{myWriteSp}{vector of TAXON_IDs provided if only subset of species
rasters are required as output.}

\item{myCropRasters}{object containing cell indices and values of input rasters cropped to area of interest created by cropToOutput()}
}
\value{
list of two or three data frames:
\itemize{
\item SpYearSummWide summary of relative abundance of species by pivoted wide by SEASONS
\item SpYearSummLong Long Format summary of relative abundance of species by SEASONS
\item grpSpYearSumm summary of abundance of species by pivoted wide by SEASONS grouped by
myAllCombs
}
}
\description{
Species' relative abundance calculation and summary
}
\details{
Calculates the relative abundance of species for each raster cell in
analysis and summaries these as summed abundance each season. Optionally it
also write relative abundance rasters for species to disk
}
