% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_TFI_2.R
\name{calc_TFI_2}
\alias{calc_TFI_2}
\title{Main Tolerable fire interval (TFI) status calculation}
\usage{
calc_TFI_2(
  myFHAnalysis = FHAnalysis,
  myCropRasters = cropRasters,
  myAllCombs = allCombs,
  myTFI_LUT = TFI_LUT,
  OutputRasters = makeTFIRasters,
  myResultsDir = ResultsDir
)
}
\arguments{
\item{myFHAnalysis}{list containing all the fire history spatial attributes created by function fhProcess()}

\item{myCropRasters}{object containing cell indices and values of input rasters cropped to area of interest created by cropToOutput()}

\item{myAllCombs}{list made by function calc_U_AllCombs}

\item{myTFI_LUT}{data.frame Lookup table from EFG for
"MIN_LO_TFI","MIN_HI_TFI","MAX_TFI","EFG_NAME", read from settings}

\item{OutputRasters}{logical whether to output rasters of TFI status for each year}

\item{myResultsDir}{path of directory where results will be written usually
generated  by FAME script}
}
\value{
data.frame with area in each TFI status for each combination in myAllCombs
}
\description{
Main Tolerable fire interval (TFI) status calculation
}
\details{
Calculates where each cell is currently at below MinTFI or above
MAX_TFI returns the per cell and long table summarised by multiple admin
units and evc
}
