% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBBTFI_2.R
\name{calcBBTFI_2}
\alias{calcBBTFI_2}
\title{Calculate area BBTFI and BBTFI rasters}
\usage{
calcBBTFI_2(
  myFHAnalysis = FHAnalysis,
  myAllCombs = allCombs,
  myCropRasters = cropRasters,
  makeBBTFIrasters = makeBBTFIrasters,
  myResultsDir = NULL
)
}
\arguments{
\item{myFHAnalysis}{list of Fhire history analysis components
created by function fhProcess()}

\item{myAllCombs}{list made by function calc_U_AllCombs}

\item{myCropRasters}{cropRasters object made using function cropToOutput.
It contains a definition of a terra::rast, cropped to the extent of the
area of interest and the values from the input rasters for that same
cropped area}

\item{makeBBTFIrasters}{logical whether or not to export
rasters for BBTFI to disk}

\item{myResultsDir}{path of directory where results will be written usually
generated  by FAME script}
}
\value{
list containing:
\itemize{
\item BBTFI_WIDE wide by SEASON  table of the number of times BBTTFI and area for each unique combination of fire history FireType,EFG, PU, and administrative subunits (District, Region etc) of area.
\item BBTFI_LONG long format table ( ie not spread by season) otherwise as BBTFI used for production of charts
}
}
\description{
Calculate area BBTFI and BBTFI rasters
}
\details{
Calculate summary  area burned below TFI BBTFI for each SEASON in analysis
(accommodating  Hi and Lo fire intensity of first burn to determine TFI)
and cumulative area BBTFI.  Also optionally outputs rasters mapping areas BBTFI
}
