% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDeltaAbund.R
\name{calcDeltaAbund}
\alias{calcDeltaAbund}
\title{Summary of changes in relative abundance}
\usage{
calcDeltaAbund(
  SpYearSumm = SpYearSumm$SpYearSummWide,
  myFHAnalysis,
  myBaseline
)
}
\arguments{
\item{SpYearSumm}{wide format data.frame output by function calc_SpeciesRA() ( one of several
data.frames in the list returned)}

\item{myFHAnalysis}{list containing all the fire history spatial attributes created by function fhProcess}

\item{myBaseline}{integer single SEASON or sequence of SEASONS used to create the baseline relative species abundance for comparison of change
#'}
}
\value{
data frame wide format summary change in relative abundance of species by SEASON relative to a Baseline
}
\description{
Summary of changes in relative abundance
}
\details{
Calculates the change in relative abundance compared to a baseline SEASON or mean of SEASONS
}
