% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_U_AllCombs.r
\name{calcU_All_Combs}
\alias{calcU_All_Combs}
\title{Calculate all combinations of input raster values}
\usage{
calcU_All_Combs(
  myFHAnalysis = FHAnalysis,
  myCropRasters = cropRasters,
  myRasterRes = RasterRes,
  puPoly = NULL
)
}
\arguments{
\item{myFHAnalysis}{list containing all the fire history spatial attributes created by function fhProcess()}

\item{myCropRasters}{list of rasters and indices and cell values created by function cropToOutput()}

\item{myRasterRes}{Resolution of Rasters to be used in the analysis set in settings file in R-script version needs to be set in shiny version}

\item{puPoly}{either NULL if no planning unit polygons are to be included in the combinations or path to planning unit polygon shapefile covering the entire extent in VicGrid94 projection}
}
\value{
list of:
\itemize{
\item data.table giving all combinations of cell values from
the input rasters for the FAME analysis
\item integer index mapping unique combinations (above) to raster cells
}
}
\description{
Calculate all combinations of input raster values
}
