% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{cellsToHectares}
\alias{cellsToHectares}
\title{Calculates multiplier to convert from raster cell count to area in hectares}
\usage{
cellsToHectares(RasterMetres = RasterRes)
}
\arguments{
\item{RasterMetres}{numeric Value cell resolution in Metres (usually from
RasterRes in settings file).}
}
\value{
numeric Multiplier to convert cell count to area in hectares
}
\description{
Calculates multiplier to convert from raster cell count to area in hectares
}
