% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSOFunc_CIM.R
\name{CIM}
\alias{CIM}
\title{Calculate mean and 95\% credible interval}
\usage{
CIM(x)
}
\arguments{
\item{x}{a vector}
}
\value{
returns the mean of x, lower and upper bounds of the 95\% credible interval of x
}
\description{
Calculate mean and 95\% credible interval
}
\details{
Internal GSO function. Used within OptRunCI()
Handy function to calculate mean and 95\% credible interval of input
}
