% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSOFunc_const_vector.R
\name{const_vector}
\alias{const_vector}
\title{Construct a n-length vector (c(k, ..., k))}
\usage{
const_vector(n, k)
}
\arguments{
\item{n}{length of vector to create}

\item{k}{the value of the units within the n length vector}
}
\value{
returns a vector
}
\description{
Construct a n-length vector (c(k, ..., k))
}
\details{
Internal GSO function. Used within props() and gso()

A handy helper function, constructs an n length vector, c(k, ... , k)
}
