% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSOFunc_ConvertExpert.R
\name{ConvertExpert}
\alias{ConvertExpert}
\title{Convert from old to new expert opinions}
\usage{
ConvertExpert(Code, x, Update)
}
\arguments{
\item{Code}{TaxonID for spp. to update}

\item{x}{value input determining how to select updates}

\item{Update}{data.frame of expert scores from ./ExpertEstimate.csv}
}
\value{
returns a converted value
}
\description{
Convert from old to new expert opinions
}
\details{
Non-internal GSO function.
Used to convert from old to new expert opinions.
}
