% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSOFunc_DataGen.R
\name{DataGen}
\alias{DataGen}
\title{Generate Growth Stage Optimisation Scores}
\usage{
DataGen(
  data,
  efgs,
  area = GSOArea,
  rule = "Rule0",
  FT = "High",
  Wt = 0.5,
  Rand = FALSE,
  NRep = 20
)
}
\arguments{
\item{data}{data.frame input called in OptRunCI}

\item{efgs}{efgs input called in OptRunCI}

\item{area}{GSO area}

\item{rule}{string value of internal GSO rule function to use}

\item{FT}{fire type string, used as a filter in 'FireType' column of data input}

\item{Wt}{rule weighting if a RUle 2 is to be used}

\item{Rand}{TRUE/FALSE to random sample for bootstrapping}

\item{NRep}{number of repetitions of bootstrapping, with replacement}
}
\value{
returns a list of scores
}
\description{
Generate Growth Stage Optimisation Scores
}
\details{
Internal GSO function. Generates the scores to use in the Growth Stage Optimisation
function (OptRunCI).  Includes random selection (Rand) for bootstrapping (NRep).
}
