% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSOFunc_eval_cs.R
\name{eval_cs}
\alias{eval_cs}
\title{Inequality contraints replacement}
\usage{
eval_cs(x, spp)
}
\arguments{
\item{x}{a list/column, e.g. run$solution,}

\item{spp}{a data.frame, usually data or option data.frame}
}
\value{
returns a vector
}
\description{
Inequality contraints replacement
}
\details{
Internal GSO function. Used within gso(), an input for nloptr()
Replaces the proportion equality constraint with <=

Since the objective is monotone, we can replace the proportion
equality constraint with <=.
Per-species limits should be do-able in the same form.
}
