% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhCheck.R
\name{fhCheck}
\alias{fhCheck}
\title{Basic Checks of input Fire History}
\usage{
fhCheck(inFH = inFH, inFHLayer = inFHLayer, validFIRETYPE = validFIRETYPE)
}
\arguments{
\item{inFH}{and input fire history polygon/mutipolygon data set with columns
SEASON and FIRETYPE  provided as shapefile, geopackage or ESRI geodatabase
file, or the name of an sf  object. if the geometries contain other than
polygon/mutipolygon an error will result.}

\item{inFHLayer}{Layer name if inFH has more than one layer
( for instance in a .gpkg) this allows selection of a particular layer,
otherwise first layer is used (Default = NULL)}

\item{validFIRETYPE}{vector of valid names in the input FIRETYPE column in
the input fire history dataset(s), if the column contains NA or values not
on this list an error will occur}
}
\value{
correctly formatted Fire History Polygon dataset as sf for use in
fhProcess1()
}
\description{
Perfoms checks on input dataset to check it is correctly formatted for use in
FAME analysis
}
