% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhProcess2.R
\name{fhProcess2}
\alias{fhProcess2}
\title{Second stage of Fire history processing for FAME}
\usage{
fhProcess2(inFH1, start.SEASON = NULL, end.SEASON = NULL, max_interval = 0)
}
\arguments{
\item{inFH1}{input partially processed Fire History from}

\item{start.SEASON}{first season for which output is wanted
(four digit year as integer), if NUll then second season in in history is used
(cannot use first season because it has no interval, this may still fail if
there is no overlap)}

\item{end.SEASON}{last season required, if NULL then maximum value in fire
history scenario used}

\item{max_interval}{maximum inter fire interval after a HIGH fire for which
subsequent fire is reassigned to HIGH, if 0 then no reassignment)}
}
\value{
A list containing: \itemize{
\item OutDF A completed Fire Analysis Simple feature collection of polygons and/or
multipolygons  with columns for Fire Season sequences, inter-fire intervals
Last fire type,last fire SEASON and last burned SEASON for each season selected
in the input range.
\item TimeSpan integer vector sequence of SEASONS to include in the analysis
output
\item YSFNames names of TSF years in output, needed by downstream functions
\item LBYNames names of LBY years in output, needed by downstream functions
\item LFTNames names of LBY years in output, needed by downstream functions }
}
\description{
Takes the results of the main FH process function  and calculates
Time Since fire (TSF) and Last Fire Type (LFT) and Last burnt year (LBY)
for each SEASON as defined in the input arguments, these values are appended
to the output sf polygon dataframe.
}
\examples{
# randomFH<-generate_random_fire_history(20)
# plot(randomFH)
# outFH1<-fhProcess1(randomFH)
# plot(outFH1)
# #complete all fields in FH analysis
# outFH2<-fhProcess2(outFH1)
#
# plot(outFH2$OutDF,max.plot = 20)


}
