% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSOFunc_geomean-fun.R
\name{geomean.fun}
\alias{geomean.fun}
\title{Calculate 'geomean'}
\usage{
geomean.fun(x, spp)
}
\arguments{
\item{x}{a list/column, e.g. run$solution,}

\item{spp}{a data.frame, usually data or option data.frame}
}
\value{
returns a single value (exponential of the summed log of each row * x,
divided by no. spp)
}
\description{
Calculate 'geomean'
}
\details{
Internal GSO function. Used within Scenarios() and gso()
calculates 'geomean', the original objective. It is scaled by number of spp

The geomean objective, constraints (and corresponding gradients) all take the
same set of extra arguments.
}
