% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSOFunc_geomean-grad.R
\name{geomean.grad}
\alias{geomean.grad}
\title{Calculate geomean objective gradient}
\usage{
geomean.grad(x, spp)
}
\arguments{
\item{x}{a list/column, e.g. run$solution,}

\item{spp}{a data.frame, usually data or option data.frame}
}
\value{
returns a single value (the sum of the partial derivatives)
}
\description{
Calculate geomean objective gradient
}
\details{
Internal GSO function. Used within gso(), an input for nloptr()
calculates the sum of the partial derivatives, with respect to each variable

The reformulated objective; nloptr expects a minimization problem.
The geomean objective, constraints (and corresponding gradients) all take the
same set of extra arguments.
}
