% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSOFunc_geomean-obj.R
\name{geomean.obj}
\alias{geomean.obj}
\title{Calculate reformulated geomean objective}
\usage{
geomean.obj(x, spp)
}
\arguments{
\item{x}{a list/column, e.g. run$solution,}

\item{spp}{a data.frame, usually data or option data.frame}
}
\value{
returns a single value (the summed log of each row * x)
}
\description{
Calculate reformulated geomean objective
}
\details{
Internal GSO function. Used within gso(), an input for nloptr()
calculates 'the reformulated objective'

The reformulated objective; nloptr expects a minimization problem.
The geomean objective, constraints (and corresponding gradients) all take the
same set of extra arguments.
}
