% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_G.R
\name{geoMean}
\alias{geoMean}
\title{Calculate Geometric mean}
\usage{
geoMean(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
numeric or NA if values are not all > 0
}
\description{
Calculate Geometric mean
}
\details{
Calculates geometric mean of a vector of positive numeric values
}
