% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_Spp_No}
\alias{get_Spp_No}
\title{Extract VBA (Victorian Biodiversity Atlas)  species ID numbers from file
paths extracts four or five digit species numbers (Victorian Biodiversity
Atlas TAXON_IDs) from vector of paths or file names containing files of e.g.
species HDMS containing the 5 digit TAXON_ID in their name}
\usage{
get_Spp_No(x = "Vector of Sp file Pathnames")
}
\arguments{
\item{x}{Vector of species file Pathnames containing VBA numbers}
}
\value{
numeric vector of 4or 5 digits (ususally TAXON_ID)
}
\description{
Extract VBA (Victorian Biodiversity Atlas)  species ID numbers from file
paths extracts four or five digit species numbers (Victorian Biodiversity
Atlas TAXON_IDs) from vector of paths or file names containing files of e.g.
species HDMS containing the 5 digit TAXON_ID in their name
}
