% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputRasters.R
\name{inputRasters}
\alias{inputRasters}
\title{Set correct input general rasters}
\usage{
inputRasters(RasterRes)
}
\arguments{
\item{RasterRes}{numeric raster resolution of the analysis in metres (
usually set in settings file or shiny app)}
}
\value{
list of input raster names correct for RasterRes or error if
RasterRes is not 75 or 225
}
\description{
Set correct input general rasters
}
