% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jfmp1.R
\name{jfmp1}
\alias{jfmp1}
\title{JFMP Stage 1 calculations}
\usage{
jfmp1(
  myPUPath,
  grpSpYearSumm,
  myAllCombs,
  myTaxonList,
  myBBTFI,
  myJFMPSeason0,
  zoneWt,
  jfmpMetricWt
)
}
\arguments{
\item{myPUPath}{path to the planning unit/burn unit  shapefile covering the area.This file must contain fields:
\itemize{
\item  PU (unique integer identifier for each planning unit/burn unit)
\item  Hectares(area in Hectares),
\item  LP1_NoBurn, LP2_NoBurn, LP1_Burn, LP2_Burn - the scores for two "Life and property" metrics for each polygon in burned and unburned state at JFMPSeason0 + 4.}}

\item{grpSpYearSumm}{grpSpYearSumm summary of abundance of species by pivoted wide by SEASONS grouped by myAllCombs which is an output of function calc_SpeciesRA.}

\item{myAllCombs}{list object retuned by function calc_U_AllCombs() that contains  combinations of input raster values for the analysis}

\item{myTaxonList}{data.frame of species attributes (read from default or user provided .csv)}

\item{myBBTFI}{a list of tables produced by function calcBBTFI_2 list containing:
\itemize{
\item BBTFI_WIDE wide by SEASON  table of the number of times BBTTFI and area for each unique combination of fire history FireType,EFG, PU, and administrative subunits (District, Region etc) of area.
\item BBTFI_LONG long format table ( ie not spread by season) otherwise as BBTFI used for production of charts}}

\item{myJFMPSeason0}{4 digit integer: the fire SEASON before the first SEASON of the JFMP being processed}

\item{zoneWt}{data.frame weighting Eco vs LP for each Fire Management Zone}

\item{jfmpMetricWt}{data.frame weighting the four JFMP scores (Flora,Fauna,LP1 and LP2)}
}
\value{
dataframe of planning units/ from input shapefile attributes with appended columns for Biodiversity scores ( area first BBTFI and sum of realive abundance scores weighted by the number of pixels of each species in the study area, Differences of scores between Burned and Unburned status at JFMPseason0 +4
}
\description{
JFMP Stage 1 calculations
}
