% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jfmpBBTFISumm.R
\name{jfmpBBTFISumm}
\alias{jfmpBBTFISumm}
\title{Summarise areas never BBTFI by scenario.}
\usage{
jfmpBBTFISumm(mydraftJfmpOut)
}
\arguments{
\item{mydraftJfmpOut}{data.frame combining autoJFMP table and one
or more alternative JFMPs produced by joinDraftJFMP()}
}
\value{
data.frame summary of area never BBTFI for No Burning, AutoJFMP and each included
JFMP scenario, Total for whole area and broken down by DISTRICT_N
}
\description{
Summarise areas never BBTFI by scenario.
}
