% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jfmpRASumm.R
\name{jfmpRASumm}
\alias{jfmpRASumm}
\title{Summary of relative abundance changes comparing JFMP results}
\usage{
jfmpRASumm(
  myDraftJfmpOut,
  myGrpSpYearSummLong,
  myTaxonList,
  myStartBaseline,
  myEndBaseline,
  myJFMPSeason0
)
}
\arguments{
\item{myDraftJfmpOut}{data.frame of autoJFMP joined to alternative draft JFMPs mad by function joinDraftJFMP()}

\item{myGrpSpYearSummLong}{data.frame Long format relative abundances of species per PU and EFG and SEASON, created when calc_SpeciesRA() is run on FHAnalysis containing PU}

\item{myTaxonList}{data.frame of species attributes (read from default or
user provided .csv)}

\item{myStartBaseline}{first SEASON of years to use for calculation RA baseline}

\item{myEndBaseline}{last SEASON of years to use for calculation RA baseline}

\item{myJFMPSeason0}{the season before the commencement of the JFMP set in settings or shinyapp}
}
\value{
data.frame Summary per species for each JFMP and NoJFMP of change relative to baseline, thresholds and summed relative abundance
}
\description{
Summary of relative abundance changes comparing JFMP results
}
