% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jfmpSummary.r
\name{jfmpSummary}
\alias{jfmpSummary}
\title{Title Summary of JFMP areas burned and not burned}
\usage{
jfmpSummary(myDraftJfmpOut = "JFMPSummary with burned or unburned")
}
\arguments{
\item{myDraftJfmpOut}{JFMP table with columns for FMZ_CODE,PU,and JFMPStatus plus all the scores calculated in jfmp1()}
}
\value{
JFMP_Summary data frame CSV reporting table 2, with one row for each District and columns for:
– Hectares allocated to burns in auto-JFMP in each zone
– Total hectares allocated to burn
– Score for each metric (x4) if JFMP implemented
– Score for each metrics (x4) if JFMP not implemented
\itemize{
\item Hectares allocated in each FMZ_CODE
}
}
\description{
Title Summary of JFMP areas burned and not burned
}
