% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LBY_f.R
\name{LBY_f}
\alias{LBY_f}
\title{Calculate last burned year matrix (LBY)}
\usage{
LBY_f(M, y)
}
\arguments{
\item{M}{numeric matrix of fire sequences sequence in rows, values are SEASON}

\item{y}{numeric  SEASON}
}
\value{
matrix  of last burned year row for each unique fire history column for each SEASON
}
\description{
Calculate last burned year matrix (LBY)
}
\details{
Function to calculate last burnt year (LBY) from matrix of rows of fire season
iterating by year (y) used in calc_TFI_2
}
