% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFHProcess.r
\name{mainFHProcess}
\alias{mainFHProcess}
\title{Main Fire History Fire Sequence analysis function}
\usage{
mainFHProcess(x = inFH)
}
\arguments{
\item{x}{input Fire History sf polygons with fields SEASON and FIRETYPE}
}
\value{
output processed fire history with unique polygons each with sequence
of fires and firetypes
}
\description{
Main Fire History Fire Sequence analysis function
}
\details{
The function takes sf polygon collection of fire history
containing polygons with two fields: FIRETYPE and SEASON Where polygons of
different FIRETYPE or SEASON overlap the function constructs unique
non-overlapping polygon of their intersections ( and non intersecting areas
) and attributes each polygon with sequential fire SEASON (SEAS01, SEAS02
...) and corresponding FIRETYPE (TYPE01,TYPE02 ...)
}
\examples{
# randomFH<-generate_random_fire_history(20)
# plot(randomFH)
# #the function is run from within fhProcess1()
# outFH1<-fhProcess1(randomFH)
# plot(outFH1)
# #complete all fields in FH analysis
# outFH2<-fhProcess2(outFH1)
#
# plot(outFH2$OutDF,max.plot = 20)


}
\seealso{
\code{\link[=fhProcess1]{fhProcess1()}} mainFHProcess() normally run from within it
}
