% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_Draft_GSO_Inputs.R
\name{make_Draft_GSO_inputs}
\alias{make_Draft_GSO_inputs}
\title{Calculate the species in each EFG in given area for GSO calculations.}
\usage{
make_Draft_GSO_inputs(
  REG_NO,
  RasterRes = 225,
  PUBLIC_LAND_ONLY,
  myPoly = clipPoly,
  generalRasterDir = "./InputGeneralRasters",
  TaxonList = "D:/FAMEshiny/ReferenceTables/FAME_TAXON_LIST.csv",
  myResultsDir = ResultsDir,
  TFI_LUT = TFI_LUT
)
}
\arguments{
\item{REG_NO}{integer DELWP fire region number 1:6 ,99 for Statewide analysis, or 7 for ad hoc boundary polygon default =7 (see look up table REG_LUT for values)}

\item{RasterRes}{integer 225 - raster resolution is always 225 for this function for speed}

\item{PUBLIC_LAND_ONLY}{logical whether to restrict analysis to public land only or the whole polygon}

\item{myPoly}{default clipPoly sf polygon data frame of LF_REGIONs (default) or ad hoc polygon - used in conjunction with REG_NO}

\item{generalRasterDir}{relative path to directory containing rasters of FIRE_REG, and PUBLIC LAND (PLM_GEN)}

\item{TaxonList}{path to  species attribute table containing paths to HDMs default is "./ReferenceTables/FAME_TAXON_LIST.csv"}

\item{myResultsDir}{path of directory where output will be saved}

\item{TFI_LUT}{data.frame lookup table for EFG loaded in setup}
}
\value{
list of three data frames LMU_EFG_AREA, Spp_EFG_LMU, and LMU_Scenario used as draft inputs to aspatial GSO calculations
}
\description{
works by using the indices of the standard dimensions raster that are in the
supplied shapefile region boundary (via function cropNAborder )
}
