% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_JFMPNoBurnTab.R
\name{make_JFMPNoBurnTab}
\alias{make_JFMPNoBurnTab}
\title{Summary of changes in relative abundance}
\usage{
make_JFMPNoBurnTab(myFHAnalysis = FHAnalysis, JFMPSeason0)
}
\arguments{
\item{myFHAnalysis}{list containing all the fire history spatial attributes created by function fhProcess}

\item{JFMPSeason0}{integer single SEASON (default to current SEASON) which is the season before the first year of burns ina 3 year JFMP
#'}
}
\value{
data framethree columns "YSFNoBurn","LBYNoBurn","LFTNoBurn" where
YSFNoBurn = YSF +4 of SEASON= JFMPSeason0
}
\description{
Summary of changes in relative abundance
}
\details{
Function used in  of JFMP scores adds columns to end of FHAnalysis$OutDF for YSF scenario wehre there is no burning for four years from JFMPSseason0
}
