% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeAbundDataLong.R
\name{makeAbundDataLong}
\alias{makeAbundDataLong}
\title{Makes Long format  Fauna  abundance table}
\usage{
makeAbundDataLong(
  AbundDataByGSFile = "./ReferenceTables/OrdinalExpertLong.csv",
  myEFG_TSF_4GS = EFG_TSF_4GS
)
}
\arguments{
\item{AbundDataByGSFile}{.csv input file containing fields
"EFG_NO", "GS4_NO",  "FireType" , "Abund", "TAXON_ID"
with Abund values for' both FireTypes for each growth stage "GS4_NO"}

\item{myEFG_TSF_4GS}{table of each combination of "EFG_NO", "GS4_NO", and "YSF" generally read in at beginning of FAME in settings file}
}
\value{
long format table with one row for each combination of "EFG_NO", "GS4_NO",  "FireType" , "Abund", "TAXON_ID" and "YSF"
sorted by TAXON_ID
}
\description{
Makes Long format  Fauna  abundance table
}
\details{
Supporting function to Make long format table for fauna abundance scores
by "TAXON_ID"  ,FireType EFG and Growth Stage from input wide format table currently
deals only with FireType of "High" and "Low"
which are converted to 2 and 1 respectively
}
