% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGS_LU.R
\name{makeGS_LU}
\alias{makeGS_LU}
\title{Make long format Growth Stage Lookup matrix}
\usage{
makeGS_LU(myEFG_TSF_4GS = EFG_TSF_4GS)
}
\arguments{
\item{myEFG_TSF_4GS}{data.fame of growth stages for each EFG with start and end years}
}
\value{
matrix rows YSF, columns EFG_NO, values GS number (1:4)
}
\description{
Make long format Growth Stage Lookup matrix
}
\details{
expands a growth stage lookup table (provided in settings file) from four growth stages (1:4)
per EFG with their years since fire spans as min(YSF) and max(YSF) to an array
with YSF as row, EFG_NO as column and growth stage (1:4) as value.   NOTE:
YSF has 1 added to both the Lookup and the input to deal with YSF==0 which
cannot be used in the array indexing
}
