% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGS_Summary.R
\name{makeGS_Summary}
\alias{makeGS_Summary}
\title{Summarise area by growth stage.}
\usage{
makeGS_Summary(myFHAnalysis, myAllCombs)
}
\arguments{
\item{myFHAnalysis}{list containing all the fire history spatial attributes
created by function fhProcess}

\item{myAllCombs}{list made by function calc_U_AllCombs}
}
\value{
list of two data.frames grouped by EFG, EFG_NAME, PLM ,FIRE_FMZ_NAME,
FIRE_REGION_NAME, DELWP_REGION
\itemize{
\item GS_Summary_wide Wide format table summarises area by Growth Stage and SEASON
\item GS_Summary_long Long format table summarises area by Growth Stage and SEASON
}
}
\description{
Summarise area by growth stage.
}
\details{
Generates wide and long format summary of area for each EFG and
season grouped by EFG, EFG_NAME, PLM ,FIRE_FMZ_NAME, FIRE_REGION_NAME,
DELWP_REGION.
}
