% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{notAllIn}
\alias{notAllIn}
\title{Checks whether all values in  one vector are in another vector}
\usage{
notAllIn(x, v = V)
}
\arguments{
\item{x}{Vector of values to check if all are in second vector}

\item{v}{Second vector of values that may or may not contain all values in x}
}
\value{
logical
}
\description{
Checks whether all values in  one vector are in another vector
}
