% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSOFunc_RunOptCI.R
\name{OptRunCI}
\alias{OptRunCI}
\title{Run Credible Intervals on GSO calcs (main GSO function)}
\usage{
OptRunCI(
  data,
  efgs,
  Scen = GSOScen,
  area = GSOArea,
  rule = "Rule0",
  FiTy = "High",
  Weight = 0.5,
  NRep = 20,
  N = 10,
  Comp = Comparison
)
}
\arguments{
\item{data}{data.frame, input data}

\item{efgs}{efgs - from wider FAME?}

\item{Scen}{data.frame, input scenarios}

\item{area}{the GSO area - wider FAME?}

\item{rule}{The chosen rule to apply expert opinion and observational data. Default 'Rule0'}

\item{FiTy}{fire type string, used as a filter in 'FireType' column of data input. Default 'High'}

\item{Weight}{rule weighting if a RUle 2 is to be used. Default 0.5}

\item{NRep}{number of repetitions of bootstrapping, with replacement. Default 20}

\item{N}{number of simulations}

\item{Comp}{The comparison species.}
}
\value{
returns a list.
}
\description{
Run Credible Intervals on GSO calcs (main GSO function)
}
\details{
The main Growth Stage Optimisation (GSO) function.

The Geometric Mean ABundance (GMA) for an area can be calculated using the
relative abundance of each species across Ecological Fire Group (EFG)
Growth Stages (GS) weighted by the fraction of that GS in that EFG available
in the area of interest. Then using non-linear optimisation, the optimal
allocation of GSs within each EFG can be calculated, giving a single answer,
based on the input from the observational data and expert opinion supplied.
However, there is a level of uncertainty with this data that should be
accounted for, as those inputs are only estimates. The uncertainty of the
GSO is estimated using a bootstrapping process (repeated samples and calculating 95\% CIs)
}
