% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageForDashboard.R
\name{packageForDashboard}
\alias{packageForDashboard}
\title{packageForDashboard
takes a saved analysis file from FAME and packages it for use in FAME dashboard app}
\usage{
packageForDashboard(inPath = "path to saved analysis from fAME as .qs file")
}
\arguments{
\item{inPath}{path to saved analysis from fAME as .qs file}
}
\value{
None but saves the packaed file for use in dashboard to "./FH_Outputs/Dashboard_"
}
\description{
packageForDashboard
takes a saved analysis file from FAME and packages it for use in FAME dashboard app
}
