% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepFH.R
\name{prepFH}
\alias{prepFH}
\title{PrepareFire History}
\usage{
prepFH(inFH = myFH)
}
\arguments{
\item{inFH}{input fire history}
}
\value{
A list containing: \itemize{
\item hasData list of one or more sf Fire History sub units of the input Fire
History containing fire boundaries, defined by a grid over the bounding box
of inFH with a grid of squreside*squareside cells
\item hasIntersects a logical vector indicating which elements of "hasData"
have intersecting Fire boundaries within them
\item squareSide integer number of units of one side of the grid used to divide
the inFH to produce "hasData" for parrallel compute.
Calculated as the floor() of area of inFH bounding box/10^8
(ie a 10km square expressed in square metres)}
}
\description{
A function that handles the subdivision of a Fire History Dataset  for
Parallel processing in the FAME fire history processing workflow.
It calculates the number of grid cells to be used in parallelizing function.
splitting the input fire history into separate sf datasets on a grid with
areas of approximately 100 square km
}
