% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSpatRasterList.R
\name{readSpatRasterList}
\alias{readSpatRasterList}
\title{Read  qs file containing list with SpatRaster objects ( package terra)}
\usage{
readSpatRasterList(filePath = "filepath to save list file to ending in qs")
}
\arguments{
\item{filePath}{file path to read the list object from ( as a package qs file with extension qs)}
}
\value{
list - if the input file contained saved PackedSpatRasters the list will contain the usable SpatRasters
}
\description{
Read  qs file containing list with SpatRaster objects ( package terra)
}
\details{
Allows reloading of a list from and R session that contains SpatRasters the function recursively searches for PackedSpatRaster objects and coverts them to "SpatRaster" using terra::rast().  A work around to allow saving of sessions of R containing SpatRasters that can be reloaded ( saving the SpatRaster itself results in a null pointer error)
to save list containing SpatRasters from R use saveSpatRasterList
}
