% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{removeEmptyDirs}
\alias{removeEmptyDirs}
\title{remove empty directories from path}
\usage{
removeEmptyDirs(rootDir = "./Results")
}
\arguments{
\item{rootDir}{relative path to remove all empty subdirectories from. Default
value "./Results"}
}
\description{
remove empty directories from path
}
\details{
Removes all empty subdirectories from the nominated path does not remove the nominated path directory even if empty
}
