% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSOFunc_Rule0.R
\name{Rule0}
\alias{Rule0}
\title{Rule 0: Use mean expert opinion only}
\usage{
Rule0(data, Max = 1)
}
\arguments{
\item{data}{data.frame input called in OptRunCI (and subsequently used in DataGen)}

\item{Max}{THIS INPUT IS NOT CALLED IN FUNCTION}
}
\value{
returns a mean value for the data.frame input 'Expert' column
}
\description{
Rule 0: Use mean expert opinion only
}
\details{
Internal GSO function. Rules functions appropriately combine
observational and expert data. These rules are used in the functions that
generates scores to use in the Growth Stage Optimisation (GSO).

Rule 0: This rule uses only expert opinion, and returns the mean value
}
