% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSOFunc_Rule1.R
\name{Rule1}
\alias{Rule1}
\title{Rule1: Unweighted mean survey data, otherwise mean expert data}
\usage{
Rule1(data)
}
\arguments{
\item{data}{data.frame input called in OptRunCI (and subsequently used in DataGen)}
}
\value{
returns a mean value for the data.frame input 'Survey' column if data exists,
else uses 'Expert' column
}
\description{
Rule1: Unweighted mean survey data, otherwise mean expert data
}
\details{
Internal GSO function. Rules functions appropriately combine observational
and expert data. These rules are used in the functions that generates scores
to use in the Growth Stage Optimisation (GSO).

Rule 1: This rule uses the mean of survey data where available, otherwise uses
expert opinion. It returns the mean value.
}
