% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSOFunc_Rule1a.R
\name{Rule1a}
\alias{Rule1a}
\title{Rule1a: Max survey data where available, otherwise max expert data}
\usage{
Rule1a(data)
}
\arguments{
\item{data}{data.frame input called in OptRunCI (and subsequently used in DataGen)}
}
\value{
returns a max value for the data.frame input 'Survey' column if data exists,
else uses 'Expert' column max
}
\description{
Rule1a: Max survey data where available, otherwise max expert data
}
\details{
Internal GSO function. Rules functions appropriately combine observational
and expert data. These rules are used in the functions that generates scores
to use in the Growth Stage Optimisation (GSO).

Rule 1a: This rule uses the max of survey data where available, otherwise uses
expert opinion. It returns the max value.
}
