% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSOFunc_Rule1b.R
\name{Rule1b}
\alias{Rule1b}
\title{Rule1b: Median survey data where available, otherwise mean expert data}
\usage{
Rule1b(data)
}
\arguments{
\item{data}{data.frame input called in OptRunCI (and subsequently used in DataGen)}
}
\value{
returns a median value for the data.frame input 'Survey' column if data exists,
else uses 'Expert' column median
}
\description{
Rule1b: Median survey data where available, otherwise mean expert data
}
\details{
Internal GSO function. Rules functions appropriately combine observational
and expert data. These rules are used in the functions that generates scores
to use in the Growth Stage Optimisation (GSO).

Rule 1b: This rule uses the median of survey data where available, otherwise
uses expert opinion. It returns the median value.
}
