% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSOFunc_Rule1c.R
\name{Rule1c}
\alias{Rule1c}
\title{Rule1c: Upper quartile survey data where available, otherwise upper quartile expert data}
\usage{
Rule1c(data)
}
\arguments{
\item{data}{data.frame input called in OptRunCI (and subsequently used in DataGen)}
}
\value{
returns the upper quartile value for the data.frame input 'Survey' column if
data exists, else uses 'Expert' column upper quartile
}
\description{
Rule1c: Upper quartile survey data where available, otherwise upper quartile expert data
}
\details{
Internal GSO function. Rules functions appropriately combine observational
and expert data. These rules are used in the functions that generates scores
to use in the Growth Stage Optimisation (GSO).

Rule 1c: This rule uses the upper quartile of survey data where available, otherwise
uses expert opinion. It returns the upper quartile value.
}
