% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSOFunc_Rule2.R
\name{Rule2}
\alias{Rule2}
\title{Rule2: Weighted average of survey and expert data.}
\usage{
Rule2(data, Weight = 0.5)
}
\arguments{
\item{data}{data.frame input called in OptRunCI (and subsequently used in DataGen)}

\item{Weight}{percentage decimal weighting towards the observational averages}
}
\value{
returns a weighted average for the mean data.frame input 'Survey' and 'Expert'
column if 'Survey' data exists, else uses the mean 'Expert' column
}
\description{
Rule2: Weighted average of survey and expert data.
}
\details{
Internal GSO function. Rules functions appropriately combine observational
and expert data. These rules are used in the functions that generates scores
to use in the Growth Stage Optimisation (GSO).

Rule 2: This rule uses a weighted average of the mean of the observational data
and the expert opinion where available, else uses the expert opinion
}
