% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSOFunc_Rule2a.R
\name{Rule2a}
\alias{Rule2a}
\title{Rule2a: Weighted average of max survey and expert data.}
\usage{
Rule2a(data, Weight = 0.5)
}
\arguments{
\item{data}{data.frame input called in OptRunCI (and subsequently used in DataGen)}

\item{Weight}{percentage decimal weighting towards the observational averages}
}
\value{
returns a weighted average for the max data.frame input 'Survey' and 'Expert'
column if 'Survey' data exists, else uses the 'Expert' column max
}
\description{
Rule2a: Weighted average of max survey and expert data.
}
\details{
Internal GSO function. Rules functions appropriately combine observational
and expert data. These rules are used in the functions that generates scores
to use in the Growth Stage Optimisation (GSO).

Rule 2a: This rule uses a weighted average of the max of the observational data
and the expert opinion where available, else uses the max expert opinion
}
