% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSOFunc_Rule2b.R
\name{Rule2b}
\alias{Rule2b}
\title{Rule2b: Weighted average of median survey and expert data.}
\usage{
Rule2b(data, Weight = 0.5)
}
\arguments{
\item{data}{data.frame input called in OptRunCI (and subsequently used in DataGen)}

\item{Weight}{percentage decimal weighting towards the observational averages}
}
\value{
returns a weighted average for the median data.frame input 'Survey' and 'Expert'
column if 'Survey' data exists, else uses the 'Expert' column median
}
\description{
Rule2b: Weighted average of median survey and expert data.
}
\details{
Internal GSO function. Rules functions appropriately combine observational
and expert data. These rules are used in the functions that generates scores
to use in the Growth Stage Optimisation (GSO).

Rule 2b: This rule uses a weighted average of the median of the observational
data and the expert opinion where available, else uses the median expert opinion
}
