% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSOFunc_Rule2c.R
\name{Rule2c}
\alias{Rule2c}
\title{Rule2c: Weighted average of upper quartile survey and expert data.}
\usage{
Rule2c(data, Weight = 0.5)
}
\arguments{
\item{data}{data.frame input called in OptRunCI (and subsequently used in DataGen)}

\item{Weight}{percentage decimal weighting towards the observational averages}
}
\value{
returns a weighted average for the upper quartile data.frame input 'Survey'
and 'Expert' column if 'Survey' data exists, else uses the the 'Expert'
column upper quartile
}
\description{
Rule2c: Weighted average of upper quartile survey and expert data.
}
\details{
Internal GSO function. Rules functions appropriately combine observational
and expert data. These rules are used in the functions that generates scores
to use in the Growth Stage Optimisation (GSO).
Rule 2c: This rule uses a weighted average of the upper quartile of the observational
data and the expert opinion where available, else uses the upper quartile of
the expert opinion
}
