% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSOFunc_Rule3a.R
\name{Rule3a}
\alias{Rule3a}
\title{Rule3a: Max of survey data only}
\usage{
Rule3a(data)
}
\arguments{
\item{data}{data.frame input called in OptRunCI (and subsequently used in DataGen)}
}
\value{
returns max value for the data.frame input 'Survey' column if data exists,
else returns NA
}
\description{
Rule3a: Max of survey data only
}
\details{
Internal GSO function. Rules functions appropriately combine observational
and expert data. These rules are used in the functions that generates scores
to use in the Growth Stage Optimisation (GSO).

Rule 3a: This rule uses the maximum of where the observational data exists,
otherwise uses nothing
}
