% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSOFunc_Rule3b.R
\name{Rule3b}
\alias{Rule3b}
\title{Rule3b: Median of survey data only}
\usage{
Rule3b(data)
}
\arguments{
\item{data}{data.frame input called in OptRunCI (and subsequently used in DataGen)}
}
\value{
returns median value for the data.frame input 'Survey' column if data exists,
else returns NA
}
\description{
Rule3b: Median of survey data only
}
\details{
Internal GSO function. Rules functions appropriately combine observational
and expert data. These rules are used in the functions that generates scores
to use in the Growth Stage Optimisation (GSO).

Rule 3b: This rule uses the median of where the observational data exists,
otherwise uses nothing
}
