% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSOFunc_Rule3c.R
\name{Rule3c}
\alias{Rule3c}
\title{Rule3c: Upper quartile of survey data only}
\usage{
Rule3c(data)
}
\arguments{
\item{data}{data.frame input called in OptRunCI (and subsequently used in DataGen)}
}
\value{
returns upper quartile value for the data.frame input 'Survey' column if data exists,
else returns NA
}
\description{
Rule3c: Upper quartile of survey data only
}
\details{
Internal GSO function. Rules functions appropriately combine observational
and expert data. These rules are used in the functions that generates scores
to use in the Growth Stage Optimisation (GSO).

Rule 3c: This rule uses the upper quartile of where the observational data exists,
otherwise uses nothing
}
