% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{saveFHsettings}
\alias{saveFHsettings}
\title{Save Settings for fhProcess()}
\usage{
saveFHsettings(outputSettings, rv)

saveFHsettings(outputSettings, rv)
}
\arguments{
\item{outputSettings}{list of settings accumulated in FAME processing}

\item{rv}{reactivevalues() object including settings defined in shiny app or
list of same object made in script version}
}
\value{
outputSettings list of settings accumulated in FAME processing with
updated values

outputSettings list of settings accumulated in FAME processing with
updated values
}
\description{
utility/helper function to capture the settings used in creating the fire
history (fhProcess()) analysis part of FAME

utility/helper function to capture the settings used in creating the fire
history (fhProcess()) analysis part of FAME
}
