% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{saveJFMPsettings}
\alias{saveJFMPsettings}
\title{Save JFMP settings}
\usage{
saveJFMPsettings(outputSettings, rv)

saveJFMPsettings(outputSettings, rv)
}
\arguments{
\item{outputSettings}{list of settings accumulated in FAME processing}

\item{rv}{reactivevalues() object including settings defined in shiny app or
list of same object made in script version}
}
\value{
outputSettings list of settings accumulated in FAME processing with
updated values

outputSettings list of settings accumulated in FAME processing with
updated values
}
\description{
utility/helper function to capture the settings used in FAME analysis

utility/helper function to capture the settings used in FAME analysis
}
