% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{saveSettingsTable}
\alias{saveSettingsTable}
\title{Save settings to csv file}
\usage{
saveSettingsTable(outputSettings, resultsDir, outputNames)

saveSettingsTable(outputSettings, resultsDir, outputNames)
}
\arguments{
\item{outputSettings}{list of settings accumulated in FAME processing}

\item{resultsDir}{the results directory where the ouptutSettings.csv file is saved}

\item{outputNames}{data.frame formatted and named as the settingstable will be output, but with column 2 values as NA}
}
\value{
writes a table as csv file to resultsDir

writes a table as csv file to resultsDir
}
\description{
untility helper function for FAME to make easily readable file of settings used
in analysis,stored with results output.

utility helper function for FAME to make easily readable file of settings used
in analysis,stored with results output.
}
