% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSOFunc_Scenarios.R
\name{Scenarios}
\alias{Scenarios}
\title{Joins GSO scenarios to input data}
\usage{
Scenarios(data, Scen = GSOScen, efg = UsedEFG)
}
\arguments{
\item{data}{data.frame, input data}

\item{Scen}{data.frame, input scenarios}

\item{efg}{ecological fire group.  THIS INPUT IS NOT CALLED IN FUNCTION}
}
\value{
returns a data.frame containing (left) join of 'data' and 'Scen' inputs
}
\description{
Joins GSO scenarios to input data
}
\details{
Internal GSO function.

Joins GSO scenarios to input data; prepares data to run models on scenarios
}
