% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectFileModule.r
\name{selectFileServer}
\alias{selectFileServer}
\title{Module Server function for file selection}
\usage{
selectFileServer(id, root_dirs = c(root = "."), filetypes = NULL)
}
\arguments{
\item{id}{string namespace id for module (same as in server)}

\item{root_dirs}{named list of paths that for the root directories within
which files may be selected default is the cwd ie c(root = "."))}

\item{filetypes}{vector of filetypes (by file extension) that may be selected from
eg c("txt", "csv", "gpkg")}
}
\value{
filepath as reactiveValue named same as module id  ( relative to root_dirs argument in fileSelectServer)
}
\description{
Module Server function for file selection
}
\details{
Select File Module UI used with selectFileUI
modified from https://stackoverflow.com/questions/38747129/how-to-use-shinyfiles-package-within-shiny-modules-namespace-issue
}
\examples{
#simple file selection shiny app
library(shiny)
ui <-   fluidPage(
 selectFileUI(
   id = "fileSelect1",label ="Select File"
 ),
 textOutput("datapath")

)

server <- function(input, output, session) {
 rv<-reactiveValues()
 x <- selectFileServer(id = "fileSelect1")
 observe(rv$datapath<-x$datapath)
 observe(output$datapath<-renderText(rv$datapath))


}
shinyApp(ui,server)
}
