% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSOFunc_SppRes.R
\name{SppRes}
\alias{SppRes}
\title{Collate species results}
\usage{
SppRes(OptDat, OptRes, Scen = GSOScen, TArea = GSOArea)
}
\arguments{
\item{OptDat}{a list, the results of the DataGen() function}

\item{OptRes}{a data.frame, the results of the gso() function}

\item{Scen}{the GSO scenario (GSOScen)}

\item{TArea}{GSO area (GSOArea)}
}
\value{
returns a data.frame
}
\description{
Collate species results
}
\details{
Internal GSO function. Used within OptRunCI()

A function to collate results for scenarios
}
