% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{unlistPivot_wider}
\alias{unlistPivot_wider}
\title{Fix Pivot_wider list of lists columns}
\usage{
unlistPivot_wider(df)
}
\arguments{
\item{df}{wide format data frame with fields that are lists of lists}
}
\value{
wide format data frame without fields that are lists of lists
}
\description{
Fix Pivot_wider list of lists columns
}
\details{
Supporting function to deal with pivot_wider returning list of lists in some cases
}
