% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileUploadModule.R
\name{uploadFileServer}
\alias{uploadFileServer}
\title{uploadFile Server part of uploadFle shiny module}
\usage{
uploadFileServer(
  id,
  filetype = c(NULL, "geog", "any other file suffix can be used ( without the dot)"),
  saveToPath = "path to save files to on server"
)
}
\arguments{
\item{id}{module id name}

\item{filetype}{one of three alternatives:
"all" which allows any selected file(s) to be uploaded
"geog" which allows load of either a geopackage (.gpkg) file OR
the four critical components of a shapefile ( .shp, .prj, .shx,.dbf) OR
a zip file containing one or more shapefiles, geopackages or
esri geodatabases. Thee zip file is unzipped after upload. OR
any other specified file suffix - tests for and allows upload
only of the defined suffix}

\item{saveToPath}{Path to save files to on server}
}
\value{
text message on status of upload for printing to ui
}
\description{
uploadFile Server part of uploadFle shiny module
}
\examples{
#app to upload files into cwd/temp

if(!dir.exists("temp")){dir.create("temp")}

ui <- fluidPage(
uploadFileUI(id = "myID",
label = "Select files",
multiple = T)
)

server <- function(input, output) {

uploadFileServer(id = "myID",
saveToPath = "./temp",
filetype ="geog")

}
shiny::shinyApp(ui = ui, server = server)
}
