% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileUploadModule.R
\name{uploadFileUI}
\alias{uploadFileUI}
\title{uploadFile UI part of uploadFle shiny module}
\usage{
uploadFileUI(id, label = "Select file", multiple = TRUE)
}
\arguments{
\item{id}{module id name}

\item{label}{label for shiny fileInput button}

\item{multiple}{default TRUE allows upload of multiple
files depending on filetype argument in server}
}
\value{
dataframe of files fed to server fileCopy
}
\description{
uploadFile UI part of uploadFle shiny module
}
\examples{
#app to upload files into cwd/temp

if(!dir.exists("temp")){dir.create("temp")}

ui <- fluidPage(
uploadFileUI(id = "myID",
label = "Select files",
multiple = T)
)

server <- function(input, output) {

uploadFileServer(id = "myID",
saveToPath = "./temp",
filetype ="geog")

}
shiny::shinyApp(ui = ui, server = server)
}
